+++ src/Process/BaseProcess.php
@@ -14,52 +14,9 @@
 declare(strict_types = 1);
 namespace Origin\Process;
 
-use RuntimeException;
-
 abstract class BaseProcess
 {
-    /**
-     * @var array
-     */
-    protected $env = [];
-
-    /**
-     * @var string
-     */
-    protected $directory;
-
-    /**
-     * @var string
-     */
-    protected $command;
-
-    /**
-     * @var boolean
-     */
-    protected $outputEnabled = false;
-
-    /**
-     * @param string $directory
-     * @return void
-     */
-    protected function setDirectory(string $directory): void
-    {
-        if (! is_dir($directory)) {
-            throw new RuntimeException('Invalid directory');
-        }
-        $this->directory = $directory;
-    }
-
-    /**
-     * @param array $env
-     * @return void
-     */
-    protected function setEnv(array $env): void
-    {
-        foreach ($env as $key => $value) {
-            $this->env[$key] = $value;
-        }
-    }
+  
     /**
      * @return boolean
      */
@@ -67,59 +24,7 @@ abstract class BaseProcess
     {
         return function_exists('posix_isatty') && posix_isatty(STDOUT);
     }
-
-    /**
-     * Escapes a command for use
-     *
-     * @param string|array $stringOrArray
-     * @return string|array
-     */
-    private function escapeCommand($stringOrArray)
-    {
-        if (is_string($stringOrArray)) {
-            return escapeshellcmd($stringOrArray);
-        }
-
-        return array_map('escapeshellarg', $stringOrArray);
-    }
-
-    /**
-     * @param string|array $stringOrArray
-     * @param boolean $escape
-     * @return void
-     */
-    protected function setCommand($stringOrArray, bool $escape = true): void
-    {
-        if ($escape) {
-            $stringOrArray = $this->escapeCommand($stringOrArray);
-        }
-
-        $this->command = is_array($stringOrArray) ? implode(' ', $stringOrArray) : $stringOrArray;
-    }
-
-    /**
-     * @see https://www.php.net/manual/en/function.proc-open.php
-     *
-     * @param boolean $output
-     * @return array
-     */
-    protected function descriptorspec(): array
-    {
-        if ($this->outputEnabled && $this->isTTY()) {
-            return [
-                ['file', '/dev/tty', 'r'],
-                ['file', '/dev/tty', 'w'],
-                ['file', '/dev/tty', 'w'],
-            ];
-        }
-
-        return [
-            ['pipe','r'],
-            ['pipe','w'],
-            ['pipe','w']
-        ];
-    }
-
+    
     /**
      * Checks if the process ended successfully
      *